apply review fixes to persisted operations
Keep the resource in state when a client exists with zero operations so
imports of empty clients work and full out-of-band deletion reconciles
in place. Document that publishing is not transactional on conflicts.
Move the client api wrappers to their own file, simplify the Read
reconciliation, adopt the shared namespace default helper in existing
data sources, and dedupe a test fixture.
