+++ internal/service/feature-subgraph/data_source_cosmo_feature_subgraph.go
@@ -120,10 +120,7 @@ func (d *FeatureSubgraphDataSource) Read(ctx context.Context, req datasource.Rea
 		return
 	}
 
-	namespace := data.Namespace.ValueString()
-	if namespace == "" {
-		namespace = "default"
-	}
+	namespace := utils.NamespaceOrDefault(data.Namespace).ValueString()
 
 	subgraph, apiErr := d.client.GetSubgraph(ctx, data.Name.ValueString(), namespace)
 	if apiErr != nil {
