+++ internal/service/persisted-operations/resource_cosmo_persisted_operations.go
@@ -52,7 +52,7 @@ Manages the full set of persisted operations (safelisted GraphQL operations) reg
 
 This resource owns every persisted operation of the given client: operations pushed outside of Terraform (e.g. via ` + "`wgc operations push`" + `) to the same client will show up as drift and be removed on the next apply. Destroying the resource deletes the operations it manages but leaves the client record itself in place.
 
-Operations are content-addressed: their identifier is the hex-encoded SHA-256 hash of their contents, matching the behavior of ` + "`wgc operations push`" + `.
+Operations are content-addressed: their identifier is the hex-encoded SHA-256 hash of their contents, matching the behavior of ` + "`wgc operations push`" + `. Publishing is not transactional: if an operation conflicts with an existing one (same id, different contents), the create or update fails but the non-conflicting operations remain registered; re-applying after fixing the conflict converges, since publishing is idempotent.
 
 Existing operations can be imported with the id format ` + "`federated_graph_name:namespace:client_name`" + `.
 
@@ -208,11 +208,8 @@ func (r *PersistedOperationsResource) Read(ctx context.Context, req resource.Rea
 		return
 	}
 
-	if len(remoteOperations) == 0 {
-		resp.State.RemoveResource(ctx)
-		return
-	}
-
+	// An empty remote set is drift, not deletion: the client record is the
+	// remote object this resource tracks, and its absence is handled above.
 	remoteById := make(map[string]*platformv1.GetPersistedOperationsResponse_Operation, len(remoteOperations))
 	for _, operation := range remoteOperations {
 		remoteById[operation.Id] = operation
@@ -231,18 +228,15 @@ func (r *PersistedOperationsResource) Read(ctx context.Context, req resource.Rea
 	// server-side normalization of contents does not cause spurious diffs) and
 	// surface out-of-band operations under their id so the plan shows their removal.
 	operations := make(map[string]string, len(remoteById))
-	matched := make(map[string]bool, len(stateOperations))
 	for label, contents := range stateOperations {
 		id := operationID(contents)
 		if _, ok := remoteById[id]; ok {
 			operations[label] = contents
-			matched[id] = true
+			delete(remoteById, id)
 		}
 	}
 	for id, operation := range remoteById {
-		if !matched[id] {
-			operations[id] = operation.Contents
-		}
+		operations[id] = operation.Contents
 	}
 
 	operationsValue, diags := types.MapValueFrom(ctx, types.StringType, operations)
